[modify] deletePost

+++ src/main/java/com/example/demo/src/post/PostService.java
@@ -68,17 +68,13 @@ public class PostService {
     }
 
     // 게시물 삭제
-    public void deletePost(int postIdx) throws BaseException {
-//        if(postProvider.checkUserExist(userIdx) ==0){
-//            throw new BaseException(USERS_EMPTY_USER_ID);
-//        }
+    public void deletePost(int userIdx, int postIdx) throws BaseException {
         if(postProvider.checkPostExist(postIdx) ==0){
             throw new BaseException(POSTS_EMPTY_POST_ID);
         }
-
-//        if(postProvider.checkUserPostExist(userIdx, postIdx)==0){
-//            throw new BaseException(POSTS_EMPTY_USER_POST);
-//        }
+        if(postProvider.checkUserPostExist(userIdx, postIdx)==0){
+            throw new BaseException(POSTS_EMPTY_USER_POST);
+        }
         try{
             int result = postDao.deletePost(postIdx);
             if(result == 0){
